+++ ContentPatcher/Framework/Patches/Patch.cs
@@ -6,7 +6,6 @@ using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.ConfigModels;
 using ContentPatcher.Framework.Tokens;
 using Microsoft.Xna.Framework;
-using Pathoschild.Stardew.Common;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Patches
@@ -33,7 +32,10 @@ namespace ContentPatcher.Framework.Patches
         private bool FromAssetExistsImpl;
 
         /// <summary>The <see cref="RawFromAsset"/> with support for managing its state.</summary>
-        protected IManagedTokenString RawManagedTargetAsset { get; }
+        private IManagedTokenString ManagedRawFromAsset { get; }
+
+        /// <summary>The <see cref="RawTargetAsset"/> with support for managing its state.</summary>
+        protected IManagedTokenString ManagedRawTargetAsset { get; }
 
 
         /*********
@@ -58,13 +60,13 @@ namespace ContentPatcher.Framework.Patches
         public string FromAsset { get; private set; }
 
         /// <summary>The raw asset key from which to load the local asset (if applicable), including tokens.</summary>
-        public ITokenString RawFromAsset { get; }
+        public ITokenString RawFromAsset => this.ManagedRawFromAsset;
 
         /// <summary>The normalized asset name to intercept.</summary>
         public string TargetAsset { get; private set; }
 
         /// <summary>The raw asset name to intercept, including tokens.</summary>
-        public ITokenString RawTargetAsset => this.RawManagedTargetAsset;
+        public ITokenString RawTargetAsset => this.ManagedRawTargetAsset;
 
         /// <summary>The conditions which determine whether this patch should be applied.</summary>
         public Condition[] Conditions { get; }
@@ -84,43 +86,35 @@ namespace ContentPatcher.Framework.Patches
             // reset
             bool wasReady = this.IsReady;
             this.State.Reset();
-            bool changed;
-
-            // update target asset
-            changed = this.RawManagedTargetAsset.UpdateContext(context);
-            this.TargetAsset = this.RawTargetAsset.IsReady ? this.NormalizeAssetNameImpl(this.RawTargetAsset.Value) : "";
+            bool isReady = true;
 
             // update local tokens
+            // (FromFile and Target may reference each other, so they need to be updated in a
+            // specific order. A circular reference isn't possible since that's checked when the
+            // patch is loaded.)
             this.PrivateContext.Update(context);
-            if (this.RawTargetAsset.IsReady)
-            {
-                string path = PathUtilities.NormalizePathSeparators(this.RawTargetAsset.Value);
-
-                this.PrivateContext.SetLocalValue(ConditionType.Target.ToString(), path);
-                this.PrivateContext.SetLocalValue(ConditionType.TargetWithoutPath.ToString(), Path.GetFileName(path));
-            }
+            bool changed = false;
+            if (this.ManagedRawTargetAsset.UsesTokens(ConditionType.FromFile))
+                changed |= this.UpdateFromFile(context) | this.UpdateTargetPath(context);
+            else
+                changed |= this.UpdateTargetPath(context) | this.UpdateFromFile(context);
+            isReady &= this.RawTargetAsset.IsReady && this.RawFromAsset?.IsReady != false;
 
             // update contextuals
-            changed = this.Contextuals.UpdateContext(this.PrivateContext) || changed;
+            changed |= this.Contextuals.UpdateContext(this.PrivateContext);
+            isReady &= this.Contextuals.IsReady && (!this.Conditions.Any() || this.Conditions.All(p => p.IsMatch(this.PrivateContext)));
             this.FromAssetExistsImpl = false;
 
-            // update from asset
-            this.FromAsset = this.RawFromAsset?.IsReady == true
-                ? this.NormalizeLocalAssetPath(this.RawFromAsset.Value, logName: $"{nameof(PatchConfig.FromFile)} field")
-                : null;
-            if (this.Contextuals.IsReady && this.FromAsset != null)
+            // check from asset existence
+            if (isReady && this.FromAsset != null)
             {
                 this.FromAssetExistsImpl = this.ContentPack.HasFile(this.FromAsset);
                 if (!this.FromAssetExistsImpl && this.Conditions.All(p => p.IsMatch(context)))
                     this.State.AddErrors($"{nameof(PatchConfig.FromFile)} '{this.FromAsset}' does not exist");
             }
 
-            // update ready flag
-            // note: from file asset existence deliberately isn't checked here, so we can show warnings at runtime instead.
-            this.IsReady =
-                this.Contextuals.IsReady
-                && (!this.Conditions.Any() || this.Conditions.All(p => p.IsMatch(this.PrivateContext)));
-
+            // update
+            this.IsReady = isReady;
             return changed || this.IsReady != wasReady;
         }
 
@@ -187,22 +181,77 @@ namespace ContentPatcher.Framework.Patches
             this.LogName = logName;
             this.Type = type;
             this.ContentPack = contentPack;
-            this.RawManagedTargetAsset = assetName;
+            this.ManagedRawTargetAsset = assetName;
             this.Conditions = conditions.ToArray();
             this.NormalizeAssetNameImpl = normalizeAssetName;
             this.PrivateContext = new LocalContext(scope: this.ContentPack.Manifest.UniqueID);
-            this.RawFromAsset = fromAsset;
+            this.ManagedRawFromAsset = fromAsset;
 
             this.Contextuals
                 .Add(this.Conditions)
-                .Add(assetName)
-                .Add(fromAsset);
+                .Add(assetName);
+        }
+
+        /// <summary>Try to read a tokenized rectangle.</summary>
+        /// <param name="tokenArea">The tokenized rectangle to parse.</param>
+        /// <param name="defaultX">The X value if the input area is null.</param>
+        /// <param name="defaultY">The Y value if the input area is null.</param>
+        /// <param name="defaultWidth">The width if the input area is null.</param>
+        /// <param name="defaultHeight">The height if the input area is null.</param>
+        /// <param name="area">The parsed rectangle.</param>
+        /// <param name="error">The error phrase indicating why parsing failed, if applicable.</param>
+        /// <returns>Returns whether the rectangle was successfully parsed.</returns>
+        protected bool TryReadArea(TokenRectangle tokenArea, int defaultX, int defaultY, int defaultWidth, int defaultHeight, out Rectangle area, out string error)
+        {
+            if (tokenArea != null)
+                return tokenArea.TryGetRectangle(out area, out error);
+
+            area = new Rectangle(defaultX, defaultY, defaultWidth, defaultHeight);
+            error = null;
+            return true;
+        }
+
+        /// <summary>Update the target path, and add the relevant tokens to the patch context.</summary>
+        /// <param name="context">The context for which to update.</param>
+        /// <returns>Returns whether the field changed.</returns>
+        private bool UpdateTargetPath(IContext context)
+        {
+            bool changed = this.ManagedRawTargetAsset.UpdateContext(context);
+
+            if (this.RawTargetAsset.IsReady)
+            {
+                this.TargetAsset = this.NormalizeAssetNameImpl(this.RawTargetAsset.Value);
+                this.PrivateContext.SetLocalValue(ConditionType.Target.ToString(), this.TargetAsset);
+                this.PrivateContext.SetLocalValue(ConditionType.TargetWithoutPath.ToString(), Path.GetFileName(this.TargetAsset));
+            }
+            else
+                this.TargetAsset = "";
+
+            return changed;
+        }
+
+        /// <summary>Update the 'FromFile' value, and add the relevant tokens to the patch context.</summary>
+        /// <param name="context">The context for which to update.</param>
+        /// <returns>Returns whether the field changed.</returns>
+        private bool UpdateFromFile(IContext context)
+        {
+            bool changed = this.ManagedRawFromAsset.UpdateContext(context);
+
+            if (this.RawFromAsset.IsReady)
+            {
+                this.FromAsset = this.NormalizeLocalAssetPath(this.RawFromAsset.Value, logName: $"{nameof(PatchConfig.FromFile)} field");
+                this.PrivateContext.SetLocalValue(ConditionType.FromFile.ToString(), this.FromAsset);
+            }
+            else
+                this.FromAsset = null;
+
+            return changed;
         }
 
         /// <summary>Get a normalized file path relative to the content pack folder.</summary>
         /// <param name="path">The relative asset path.</param>
         /// <param name="logName">A descriptive name for the field being normalized shown in error messages.</param>
-        protected string NormalizeLocalAssetPath(string path, string logName)
+        private string NormalizeLocalAssetPath(string path, string logName)
         {
             try
             {
@@ -225,25 +274,6 @@ namespace ContentPatcher.Framework.Patches
             {
                 throw new FormatException($"The {logName} for patch '{this.LogName}' isn't a valid asset path (current value: '{path}').", ex);
             }
-        }
-
-        /// <summary>Try to read a tokenized rectangle.</summary>
-        /// <param name="tokenArea">The tokenized rectangle to parse.</param>
-        /// <param name="defaultX">The X value if the input area is null.</param>
-        /// <param name="defaultY">The Y value if the input area is null.</param>
-        /// <param name="defaultWidth">The width if the input area is null.</param>
-        /// <param name="defaultHeight">The height if the input area is null.</param>
-        /// <param name="area">The parsed rectangle.</param>
-        /// <param name="error">The error phrase indicating why parsing failed, if applicable.</param>
-        /// <returns>Returns whether the rectangle was successfully parsed.</returns>
-        protected bool TryReadArea(TokenRectangle tokenArea, int defaultX, int defaultY, int defaultWidth, int defaultHeight, out Rectangle area, out string error)
-        {
-            if (tokenArea != null)
-                return tokenArea.TryGetRectangle(out area, out error);
-
-            area = new Rectangle(defaultX, defaultY, defaultWidth, defaultHeight);
-            error = null;
-            return true;
         }
     }
 }